Move core logic to pkg directory to have a better organization
Move core logic files to a new directory `pkg/aslanwords`.

* Add `pkg/aslanwords/generate.go` with the content from the original `generate.go`.
* Add `pkg/aslanwords/opts.go` with the content from the original `opts.go`.
* Add `pkg/aslanwords/syllables.go` with the content from the original `syllables.go`.
* Delete the original `generate.go`, `opts.go`, and `syllables.go` files from the root directory.
* Update import paths in `cmd/generate/main.go` to import from `pkg/aslanwords`.
